+++ src/apis/auth/useLogin.js
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import { axiosInstance } from "./axiosInstance";
+import { axiosInstance } from "../axiosInstance";
 import { setCookies } from "./cookie";
 
 const login = async (credentials) => {
