+++ src/apis/auth/useSignup.js
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query";
-import { axiosInstance } from "./axiosInstance";
+import { axiosInstance } from "../axiosInstance";
 
 const signup = async (credentials) => {
   const response = await axiosInstance.post("/auth/sign-up", credentials);
