+++ src/pages/Admin/Monitoring/index.jsx
@@ -13,9 +13,8 @@ import greenMarker from "@assets/images/marker-green.png";
 import greyMarker from "@assets/images/marker-grey.png";
 import redMarker from "@assets/images/marker-red.png";
 import SensorDrawer from "@components/AdminMonitoring/SensorDrawer";
-import { dummySignalLights } from "@pages/Monitoring/dummy";
 import FailureInfo from "@components/AdminMonitoring/FailureInfo";
-import { grey } from "@mui/material/colors";
+import { dummySignalLights } from "@pages/User/Monitoring/dummy";
 
 const AdminMonitoring = () => {
   return (
